Buza forms: added a urls file for Buza and migrated forms from accounts.
